Minor fixes

+++ src/clear_cache.py
@@ -5,14 +5,15 @@ import shutil
 
 def clear_cache(path):
 	for filename in os.listdir(path):
-		file_path = os.path.join(path, filename)
-		try:
-			if os.path.isfile(file_path) or os.path.islink(file_path):
-				os.unlink(file_path)
-			elif os.path.isdir(file_path):
-				shutil.rmtree(file_path)
-		except Exception as e:
-			print('WARNING: failed to delete %s. Reason: %s' % (file_path, e))
+		if filename != '.gitignore':
+			file_path = os.path.join(path, filename)
+			try:
+				if os.path.isfile(file_path) or os.path.islink(file_path):
+					os.unlink(file_path)
+				elif os.path.isdir(file_path):
+					shutil.rmtree(file_path)
+			except Exception as e:
+				print('WARNING: failed to delete %s. Reason: %s' % (file_path, e))
 	print(f'\033[95m{path:<16}\033[0mfolder successfully deleted!')
